Converted statstream to redis
There are two reasons for this,

A) It is suprisingly hard to keep pubsub internally within the
process fron hanging, for the past 6 months of this program running
I had to resort to rebooting it once a hour to make sure it didnt
hang and take out the systems that needed it with it (hey, at least
I managed to retest the reconnection logic in all of those apps)

B) Websockets are not a reasonable thing to assume all apps will
be able to use, while I can wrap things in with "webcat" to
give them websockets. I'm already running redis, and redis has
really quite good libary support. So I will just do that